Move static init scheduling to new pass
