+++ src/Makefile_obj.in
@@ -319,6 +319,7 @@ RAW_OBJS_PCH_ASTNOMT = \
   V3Reorder.o \
   V3Sampled.o \
   V3Sched.o \
+  V3SchedStatic.o \
   V3SchedAcyclic.o \
   V3SchedPartition.o \
   V3SchedReplicate.o \
