+++ src/V3SchedStatic.h
@@ -1,0 +1,30 @@
+// -*- mode: C++; c-file-style: "cc-mode" -*-
+//*************************************************************************
+// DESCRIPTION: Verilator: Post-schedule static initializer ordering
+//
+// Code available from: https://verilator.org
+//
+//*************************************************************************
+//
+// This program is free software; you can redistribute it and/or modify it
+// under the terms of either the GNU Lesser General Public License Version 3
+// or the Perl Artistic License Version 2.0.
+// SPDX-FileCopyrightText: 2026 Wilson Snyder
+// SPDX-License-Identifier: LGPL-3.0-only OR Artistic-2.0
+//
+//*************************************************************************
+
+#ifndef VERILATOR_V3SCHEDSTATIC_H_
+#define VERILATOR_V3SCHEDSTATIC_H_
+
+#include "config_build.h"
+#include "verilatedos.h"
+
+class AstNetlist;
+
+class V3SchedStatic final {
+public:
+    static void scheduleStatic(AstNetlist* netlistp) VL_MT_DISABLED;
+};
+
+#endif  // Guard
