+++ test_regress/t/t_static_init_circular.py
