+++ test_regress/t/t_static_init_class_dep.py
