+++ test_regress/t/t_static_init_good_cases.py
@@ -11,7 +11,7 @@ import vltest_bootstrap
 
 test.scenarios('simulator')
 
-test.compile(verilator_flags2=['--binary'])
+test.compile(verilator_flags2=['--binary', '--timing'])
 
 test.execute()
