+++ test_regress/t/t_uvm_static_init_min.v
@@ -1,7 +1,8 @@
 // DESCRIPTION: Verilator: Verilog Test module
 //
 // This file ONLY is placed under the Creative Commons Public Domain.
-// SPDX-FileCopyrightText: 2026 Nick
+// SPDX-FileCopyrightText: 2003-2026 Wilson Snyder
+// SPDX-FileCopyrightText: 2026 Nick Badger
 // SPDX-License-Identifier: CC0-1.0
 
 package uvm_pkg;
